continued progress on save/load
ok preliminary saveload is sort of working in the editor.
save and load buttons in file menu save only the objects in your editor, and only saves it to a single file manually for now.

obviously this is not final, but a placeholder for quicker testing / beta testers.

the format however that it's saving is feeling pretty good and well rounded.