fixed issue loading animated save file
previously, if the user saved a project that was actively playing, whatever macro's that were active at the moment, ie had a Macro fade value of > 0, had their macro fade value saved.. this meant that when the file loaded with the playhead at the start, the animation would look off, because some stray macros were contributing that maybe shouldn't have.

The solve was to forcibly set all macro fades to 0, before allowing the sequence to start playing.